+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - remove `repeatCount` prop from `@TimeSwitcher`
 - support of multiple `@TimeSwitcher` in Matlab
 - prettify code in SLV/DBSolve
+- add `groupConstBy` prop to SLV/DBSolve export
 
 ## 0.5.8